+++ CHANGELOG.md
@@ -4,6 +4,13 @@ This changelog track changes to the qoqo_iqm project starting at version 0.6.5 (
 
 ## Not released
 
+## 0.7.0
+
+* Refactored Backend to adapt to the new Resonance interface
+* Added DemoDevice and ResonatorFreeDevice
+* Removed DemoDevice
+* Updated integration tests to use Deneb
+
 ## 0.6.5
 
 * Updated to qoqo 1.5